Refactor out the access-granting function
