+++ lib/build.js
@@ -5,6 +5,7 @@ var _ = require('lodash');
 
 var github = require('./github');
 var contentService = require('./content-service');
+var grantUsersAccess = require('./access').grantUsersAccess;
 
 exports.create = function (toolbelt, repository, callback) {
   if (!repository.kind) {
@@ -264,52 +265,11 @@ var createGitHubBuild = function (toolbelt, repository, callback) {
     });
   };
 
-  var grantUsersAccess = function (callback) {
-    toolbelt.debug("Granting users access to %s.", projectName);
-
-    User.find({}, function (err, us) {
-      if (err) {
-        toolbelt.error("Unable to list users: %s", err.message);
-        return callback(null);
-      }
-
-      var grantUserAccess = function (u, cb) {
-        var alreadyHasAccess = u.projects.some(function (p) {
-          return p.name === projectName;
-        });
-
-        if (alreadyHasAccess) {
-          return cb(null);
-        }
-
-        User.update({ _id: u._id }, {
-          $push: {
-            projects: {
-              name: projectName,
-              display_name: projectName,
-              access_level: 1
-            }
-          }
-        }, function (err, count) {
-          if (err) {
-            toolbelt.error("Unable to grant user %s access to project %s", u.email, projectName);
-          } else if (count < 1) {
-            toolbelt.error("No users updated when granting user %s access to project %s", u.email, projectName);
-          }
-
-          cb(null);
-        });
-      };
-
-      async.map(us, grantUserAccess, callback);
-    });
-  };
-
   checkForExistingProject(function (err, exists) {
     if (err) return callback(null);
 
     if (exists) {
-      return grantUsersAccess(callback);
+      return grantUsersAccess(toolbelt, projectName, callback);
     }
 
     async.series([
@@ -317,7 +277,7 @@ var createGitHubBuild = function (toolbelt, repository, callback) {
       generateKeyPair,
       issueAPIKey,
       createStriderProject,
-      grantUsersAccess
+      async.apply(grantUsersAccess, toolbelt, projectName)
     ], function (err) {
       if (err) {
         toolbelt.error("Failed to create build for project %s: %s", projectName, err.message);
